Refactor to remove glob from esbuild CLI
