+++ javascript/package.json
@@ -14,12 +14,10 @@
   },
   "scripts": {
     "build": "npm run build:es && npm run build:cjs",
-    "build:cjs": "tsc && cp -r src/tests/data/ dist/tests/data/",
-    "build:es": "tsc --outDir module --module ES2020 && cp module/win-browser.js module/win.js && cp -r src/tests/data/ module/tests/data/",
+    "build:cjs": "tsc && cp -r src/tests/models/ dist/tests/models/",
+    "build:es": "tsc --outDir module --module ES2020 && cp module/win-browser.js module/win.js && cp -r src/tests/models/ module/tests/models/",
     "bundle": "npm run bundle:webcomponents && npm run bundle:test",
-    "bundle:test": "npm run bundle:test:browser && npm run bundle:test:shared",
-    "bundle:test:browser": "esbuild module/tests/browser/*.js --bundle --sourcemap --outdir=bundle/tests/browser/",
-    "bundle:test:shared": "esbuild module/tests/shared/*.js --bundle --sourcemap --outdir=bundle/tests/shared/",
+    "bundle:test": "esbuild module/tests/index.browser.js --bundle --sourcemap --outfile=bundle/tests/index.browser.js",
     "bundle:webcomponents": "npm run bundle:webcomponents:ja && npm run bundle:webcomponents:zh-hans && npm run bundle:webcomponents:zh-hant",
     "bundle:webcomponents:ja": "esbuild module/webcomponents/budoux-ja.js --bundle --minify --sourcemap --outfile=bundle/budoux-ja.min.js",
     "bundle:webcomponents:zh-hans": "esbuild module/webcomponents/budoux-zh-hans.js --bundle --minify --sourcemap --outfile=bundle/budoux-zh-hans.min.js",
@@ -31,7 +29,7 @@
     "pretest": "npm run build && npm run bundle:test",
     "test": "npm run test:jasmine && npm run test:karma && npm run test:cli-version",
     "test:cli-version": "node ./scripts/check-cli-version.js",
-    "test:jasmine": "jasmine dist/tests/shared/test_*.js dist/tests/node/test_*.js",
+    "test:jasmine": "jasmine dist/tests/index.node.js",
     "test:karma": "karma start",
     "lint": "gts lint",
     "fix": "gts fix"
